kernel: add ALL_BITS_UINT; avoid 1UL, 0UL
On some 64bit systems, "unsigned long" is actually just 32bit, so something
like `1UL << 44` won't work right (it evaluates to zero) while `(UInt)1 << 44`
will work.

+++ src/listfunc.c
@@ -1440,7 +1440,7 @@ static Obj FuncLIST_WITH_IDENTICAL_ENTRIES(Obj self, Obj n, Obj obj)
             for (; len >= BIPEB; len -= BIPEB)
                 *ptrBlist++ = ~(UInt)0;
             if (len > 0)
-                *ptrBlist |= (1UL << len) - 1;
+                *ptrBlist |= ((UInt)1 << len) - 1;
         }
     }
     else if (len == 0) {
